+++ docs/NEWS.rst
@@ -8,10 +8,15 @@ of XIST. For a description of how to update your code to each versions of XIST
 see :ref:`MIGRATION`.
 
 
-Changes in 5.20 (released ??/??/2016)
+Changes in 5.20 (released 07/29/2016)
 -------------------------------------
 
 *	Dictionaries and sets in UL4 now support the :meth:`clear` method.
+
+*	:program:`rul4` now supports saving files to disk, making log calls
+	that print message to ``stderr`` and accessing environment variables.
+	All variables passed to the templates have been moved into an object named
+	``globals``.
 
 
 Changes in 5.19.4 (released 06/30/2016)